removal of old code and comments, moving some functions into aux_func.py
